requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.8
pinecone-client==2.2.4
google-cloud-speech==2.21.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2